Events: add start/duration settings to Time Passed event
The new Start After setting delays the trigger's timer. The new Duration
setting limits the window in which the trigger will be fired. Both new
settings should be defined in number of seconds.

When the trigger's timer is done running (either because of it running
Once or when the Duration is passed), the trigger's callback is removed
from the list of timer triggers to save unnecessary function calls.