Events: fix bug in the Field Selected event
The trigger for Field Selected incorrectly fired when zero values were
selected in a field because the use of `GetFieldSelections()` returned '-'
as a string value. This was both considered not-empty for a field
selection and it can be confused with the selection of an actual field's
value of the same character.

The new implementation considers both results of `GetFieldSelections()`
and the now added request for `GetSelectedCount()`. The last function
returns an actual value of 0 when no values are selected, which can be
properly compared to any previous selections.